+++ src/main/java/com/proj/itstaym/manager/api/UserManager.java
@@ -1,17 +1,15 @@
 package com.proj.itstaym.manager.api;
 
-import java.math.BigInteger;
-import java.util.Optional;
-
+import com.proj.itstaym.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.proj.itstaym.model.User;
+import java.util.Optional;
 
 @Repository
-public interface UserManager extends JpaRepository<User, BigInteger> {
+public interface UserManager extends JpaRepository<User, Long> {
 
     @Query("SELECT u FROM User u WHERE u.email = :email")
     Optional<User> findByEmail(@Param("email") String email);
